partitions functions into module
