+++ index.html
@@ -4,8 +4,10 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>CSM | HOME</title>
+    <script type="module" src="js/module.js"></script>
+    <script type="module" src="js/script.js"></script>
   </head>
-  <body onload="onLoad()">
+  <body>
     <div id="message">
     </div>
     <div>
@@ -22,5 +24,5 @@
       </div>
     </div>
   </body>
-  <script src="js/script.js"></script>
+  <!--body onload="onLoad()">
 </html>
